Python-backend/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
joblib
flask
transformers
spacy
pdfplumber
python-docx
keybert
